+++ packages/docs/src/components/Navbar/NavbarContent/styles.module.css
@@ -1,23 +1,3 @@
-.navbar {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  gap: 24px;
-  /* Blur on/off */
-  background-color: transparent;
-  /* background-color: var(--swm-background); */
-
-  position: sticky;
-  top: 0;
-  width: 100%;
-
-  /* Override default values */
-  height: 80px;
-  padding: 0;
-
-  z-index: 1000;
-}
-
 .navbar_container {
   height: 80px;
   width: 100%;
@@ -33,6 +13,8 @@
 
   /* Blur on/off */
   backdrop-filter: blur(20px);
+
+  z-index: 1000;
 }
 
 .navbar_left {
