+++ packages/docs/src/components/Navbar/NavbarLayout/styles.module.css
@@ -1,0 +1,20 @@
+.navbar {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  gap: 24px;
+
+  /* Blur on/off */
+  background-color: transparent;
+  /* background-color: var(--swm-background); */
+
+  position: sticky;
+  top: 0;
+  width: 100%;
+
+  /* Override default values */
+  height: 80px;
+  padding: 0;
+
+  z-index: 1000;
+}
