+++ packages/docs/src/components/Navbar/NavbarLink/index.tsx
@@ -1,0 +1,15 @@
+import React from "react";
+import { NavbarItem } from "../NavbarContent";
+
+export interface NavbarLinkProps {
+  item: NavbarItem;
+  onClick?: () => void;
+}
+
+export default function NavbarLink({ item, onClick }: NavbarLinkProps) {
+  return (
+    <a href={item.to} onClick={onClick}>
+      {item.label}
+    </a>
+  );
+}
